consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["alexau <alexuau922@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
psycopg2-binary = "^2.9.9"
types-psycopg2 = "^2.9.21.19"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
coverage = "^7.3.2"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
